Actual app
Currently, there is a landing page and upon logging in you will
be able to go to (any user's) collection page. You are only able
to add things to your collection (thanks to the backend). You
can browse items to add to the collection, and you shouldn't
be able to POST a request to the backend for an instrument
that doesn't exist, thanks to the filtering that is done.